updated wii port to conform it to new devkitpro

+++ engine/wii/control.c
@@ -72,7 +72,7 @@ void reset()
 	hwbutton = WII_RESET;
 }
 
-void wiimote_poweroff(int playernum)
+void wiimote_poweroff(long int playernum)
 {
 	hwbutton = WII_SHUTDOWN;
 }
